comparison/model_segm_comparison/stardist/requirements_stardist.txt
absl-py==2.1.0
aiobotocore==2.25.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioitertools==0.13.0
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
ansys-api-fluent==0.3.36
ansys-api-platform-instancemanagement==1.1.2
ansys-api-tools-filetransfer==0.1.1
ansys-api-workbench==0.2.0
ansys-fluent-core==0.34.2
ansys-fluent-visualization==0.21.1
ansys-platform-instancemanagement==1.1.2
ansys-pythonnet==3.1.0rc6
ansys-tools-filetransfer==0.1.1
ansys-tools-path==0.7.1
ansys-units==0.7.0
ansys-workbench-core==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.5
bleach==6.2.0
botocore==1.40.61
bqplot==0.12.44
certifi==2025.1.31
cffconvert==2.0.0
cffi==1.17.1
cftime==1.6.4.post1
chaospy==4.3.21
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
clr_loader==0.2.7.post0
cmocean==4.0.3
colorama==0.4.6
colorcet==3.1.0
comm==0.2.2
contourpy==1.3.1
crc32c==2.8
csbdeep==0.8.2
cycler==0.12.1
Cython==3.2.0
debugpy==1.8.16
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
docker==7.1.0
docker-pycreds==0.4.0
docopt==0.6.2
donfig==0.8.1.post1
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.2
filelock==3.17.0
flatbuffers==25.2.10
flexcache==0.3
flexparser==0.4
fluidfoam==0.2.8
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.10.0
fvcore==0.1.5.post20221221
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
gmsh==4.13.1
gmsh_interop==2024.1
google-pasta==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-health-checking==1.62.3
grpcio-status==1.62.3
h11==0.16.0
h5py==3.12.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.2
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
importlib-resources==5.12.0
importlib_metadata==8.6.1
iopath==0.1.10
ipydatawidgets==4.3.5
ipykernel==6.30.1
ipython==8.32.0
ipyvolume==0.6.3
ipyvue==1.11.2
ipyvuetify==1.10.0
ipywebrtc==0.6.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_proxy==4.4.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.13
keras==3.8.0
kiwisolver==1.4.7
lark==1.2.2
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
meshio==5.3.5
MeshPy==2022.1.3
mistune==3.1.4
ml_dtypes==0.5.3
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.1.0
namex==0.0.8
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netCDF4==1.7.2
networkx==3.4.2
ninja==1.13.0
nltk==3.9.1
notebook==7.4.7
notebook_shim==0.2.4
numba==0.61.0
numcodecs==0.16.3
numpoly==1.3.4
numpy==2.0.2
numpy-stl==3.2.0
nvidia-physicsnemo==1.2.0
nvidia-physicsnemo.sym==2.2.0
omegaconf==2.3.0
onnx==1.19.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
OpenPIV==0.25.3
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pillow==11.0.0
Pint==0.25.2
platformdirs==4.3.6
pooch==1.8.2
portalocker==3.1.1
prometheus_client==0.22.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==4.25.6
psutil==6.1.1
pure_eval==0.2.3
pyansys-tools-report==0.8.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.18.0
pygmsh==7.1.17
pykwalify==1.8.0
pymeshio==3.0.1
pyparsing==3.2.0
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-utils==3.9.1
pythreejs==2.4.2
pytools==2025.1.1
pytorch-ranger==0.1.1
pytz==2025.1
pyvista==0.46.3
pyvistaqt==0.11.3
PyVTK==0.5.18
pywin32==308
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.1.0
QtPy==2.4.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rpds-py==0.27.1
Rtree==1.3.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3fs==2025.10.0
safetensors==0.6.2
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.1
scooby==0.10.0
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.26.1
setproctitle==1.3.5
setuptools==80.9.0
shapely==2.1.0
shellingham==1.5.4
simpervisor==1.0.0
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
stardist==0.9.2
statannotations==0.7.2
stringzilla==3.12.3
symengine==0.14.1
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensordict==0.6.1
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.18.1
tetgen==0.6.5
threadpoolctl==3.5.0
tifffile==2025.1.10
timm==1.0.22
tinycss2==1.4.0
torch==2.4.1+cu124
torch-geometric==2.6.1
torch-optimizer==0.3.0
torch-tb-profiler==0.4.3
torch_cluster==1.6.3
torch_scatter==2.1.2+pt25cu124
torch_sparse==0.6.18
torch_spline_conv==1.2.2
torchaudio==2.4.1+cu124
torchdata==0.11.0
torchvision==0.19.1+cu124
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.1
trame==3.12.0
trame-client==3.10.1
trame-common==1.0.1
trame-server==3.6.0
trame-vtk==2.9.1
trame-vuetify==3.0.2
transforms3d==0.4.2
treelib==1.8.0
trimesh==4.6.1
typer-slim==0.20.0
types-python-dateutil==2.9.0.20250822
typing==3.7.4.3
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
vtk==9.3.1
wandb==0.19.9
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
WMI==1.5.1
wrapt==1.17.2
wslink==2.4.0
xarray==2025.10.1
yacs==0.1.8
yarl==1.18.3
zarr==3.1.3
zipp==3.21.0
